graphqlで既存のcreateとindexを代替
