+++ graphql/client.ts
@@ -1,0 +1,21 @@
+import { createAuthLink, AuthOptions } from 'aws-appsync-auth-link';
+import { createSubscriptionHandshakeLink } from 'aws-appsync-subscription-link';
+import { ApolloClient, InMemoryCache, ApolloLink } from '@apollo/client/core';
+import appSyncConfig from '../aws-exports';
+
+const url = appSyncConfig.aws_appsync_graphqlEndpoint;
+const region = appSyncConfig.aws_appsync_region;
+const auth: AuthOptions = {
+  type: 'API_KEY',
+  apiKey: appSyncConfig.aws_appsync_apiKey
+};
+
+const link = ApolloLink.from([
+  createAuthLink({ url, region, auth }),
+  createSubscriptionHandshakeLink({ url, region, auth })
+]);
+
+export const client = new ApolloClient({
+  link,
+  cache: new InMemoryCache()
+});
