+++ graphql/queries/ListTodos.ts
@@ -1,0 +1,13 @@
+import { gql } from '@apollo/client';
+
+export const ListTodos = gql`
+  query ListTodos {
+    listTodos {
+      items {
+        id
+        title
+        body
+      }
+    }
+  }
+`;
